+++ packages/lib/sdk/src/plugins/datasources/evalSources.js
@@ -10,7 +10,7 @@ import { buildMultipartParquet } from '@evidence-dev/universal-sql';
 import { buildSourceDirectoryProxy } from './buildSourceDirectoryProxy.js';
 import { addToCache, checkCache, flushCache, loadCache } from './SourceResultCache.js';
 import ora from 'ora';
-import { dataUrlPrefix } from '../../build-dev/vite/virtuals/node/projectPaths.js';
+import { dataUrlPrefix } from '../../lib/projectPaths.js';
 import { subSourceVariables } from './sub-source-vars.js';
 import { logQueryEvent } from '@evidence-dev/telemetry';
 
@@ -39,14 +39,20 @@ export const evalSources = async (dataPath, metaPath, filters, strict) => {
 	pluginLoader.succeed();
 
 	/** @type {import('./types.js').Manifest} */
-	const outputManifest = { renderedFiles: {}, locatedFiles: {} };
+	const outputManifest = { renderedFiles: {}, locatedFiles: {}, locatedSchemas: [] };
+
+	/** @type {string[]} */
+	const skippedSources = [];
 
 	for (const source of sources) {
-		console.log(chalk.dim('-'.repeat(5)));
+		outputManifest.locatedSchemas ??= [];
+		outputManifest.locatedSchemas.push(source.name);
 		if (filters?.sources?.size && !filters?.sources?.has(source.name)) {
-			console.log(`  [Filtered] ${chalk.bold(source.name)}`);
+			console.debug(`  [Skipping]: ${chalk.bold(source.name)}`);
+			skippedSources.push(source.name);
 			continue;
 		} else {
+			console.log(chalk.dim('-'.repeat(5)));
 			console.log(chalk.green(`  [Processing] ${chalk.bold(source.name)}`));
 		}
 
@@ -117,11 +123,11 @@ export const evalSources = async (dataPath, metaPath, filters, strict) => {
 			outputManifest.locatedFiles[source.name].push(table.name);
 			spinner.start('Processing...');
 			if (utils.isFiltered(table.name)) {
-				spinner.warn('Skipping: Filtered');
+				spinner.info('Skipped');
 				continue;
 			}
 			if (utils.isCached(table.name, table.content)) {
-				spinner.warn('Skipping: Cached');
+				spinner.info('From Cache');
 				logQueryEvent('cache-query', source.type, source.name);
 				continue;
 			}
@@ -147,8 +153,7 @@ export const evalSources = async (dataPath, metaPath, filters, strict) => {
 					tmpDir,
 					outDir,
 					filename,
-					table.expectedRowCount,
-					1000 * 1000 // TODO: Configurable?
+					source.buildOptions.batchSize
 				);
 
 				if (writtenRows === false) {
@@ -159,6 +164,11 @@ export const evalSources = async (dataPath, metaPath, filters, strict) => {
 				outputManifest.renderedFiles[source.name].push(
 					`${dataUrlPrefix}/${source.name}/${table.name}/${filename}`
 				);
+				await fs.writeFile(
+					path.join(outDir, table.name + '.schema.json'),
+					JSON.stringify(table.columnTypes)
+				);
+
 				addToCache(source.name, table.name, table.content);
 
 				spinner.succeed(`Finished, wrote ${writtenRows} rows.`);
@@ -177,6 +187,13 @@ export const evalSources = async (dataPath, metaPath, filters, strict) => {
 		}
 	}
 	console.log(chalk.dim('-'.repeat(5)));
+
+	if (skippedSources.length)
+		console.log(
+			chalk.dim(
+				`  ${skippedSources.length} source${skippedSources.length === 1 ? '' : 's'} were not run due to filters`
+			)
+		);
 
 	await flushCache(metaPath);