+++ packages/lib/sdk/src/usql/query/Query.js
@@ -52,6 +52,7 @@ import { sterilizeQuery } from './sterilizeQuery.js';
  * @typedef {Object} QueryGlobalEvents
  * @property {undefined} inFlightQueryStart
  * @property {undefined} inFlightQueryEnd
+ * @property {{raw: Query<any>, proxied: QueryValue<any>}} queryCreated
  */
 /** @typedef {import ("../types.js").EventEmitter<QueryGlobalEvents>} QueryGlobalEventEmitter */
 
@@ -82,15 +83,22 @@ export class Query {
 	/// Data
 	/** @type {RowType[]} */
 	#data = [];
+	/** @type {number} */
+	#dataQueryTime = -1;
 	get dataLoaded() {
 		return ['resolved', 'rejected'].includes(this.#sharedDataPromise.state);
 	}
 	get dataLoading() {
 		return this.#sharedDataPromise.state === 'loading';
 	}
+	get dataQueryTime() {
+		return this.#dataQueryTime;
+	}
 	/// Length
 	/** @type {number} */
 	#length = 0;
+	/** @type {number} */
+	#lengthQueryTime = -1;
 	get length() {
 		return this.#length;
 	}
@@ -100,12 +108,17 @@ export class Query {
 	get lengthLoading() {
 		return this.#sharedLengthPromise.state === 'loading';
 	}
+	get lengthQueryTime() {
+		return this.#lengthQueryTime;
+	}
 
 	/// Columns
 	/** @type {import('../../types/duckdb-wellknown.js').DescribeResultRow[]} */
 	#columns = [];
 	/** @type {Record<keyof RowType, undefined> | undefined} */
 	#mockRow = undefined;
+	/** @type {number} */
+	#columnsQueryTime = -1;
 
 	get columns() {
 		return this.#columns;
@@ -115,6 +128,9 @@ export class Query {
 	}
 	get columnsLoading() {
 		return this.#sharedColumnsPromise.state === 'loading';
+	}
+	get columnsQueryTime() {
+		return this.#columnsQueryTime;
 	}
 
 	/**
@@ -221,7 +237,8 @@ export class Query {
 	/** @type {import("../types.js").EventMap<QueryGlobalEvents>} */
 	static #globalHandlerMap = {
 		inFlightQueryStart: new Set(),
-		inFlightQueryEnd: new Set()
+		inFlightQueryEnd: new Set(),
+		queryCreated: new Set()
 	};
 	/**
 	 * @template {keyof QueryGlobalEvents} Event
@@ -325,6 +342,9 @@ ${this.text.trim()}
 					this.#debug('long-running', `Query took ${before - after}ms to execute`);
 				}
 
+				this.#dataQueryTime = after - before;
+
+				this.#fetchLength();
 				this.#sharedDataPromise.resolve(this);
 				this.#emit('dataReady', undefined);
 				if (isPromise) {
@@ -347,9 +367,7 @@ ${this.text.trim()}
 		return resolved;
 	};
 	fetch = async () => {
-		return Promise.allSettled([this.#fetchColumns(), this.#fetchData(), this.#fetchLength()]).then(
-			() => this.value
-		);
+		return Promise.allSettled([this.#fetchColumns(), this.#fetchData()]).then(() => this.value);
 	};
 	/**
 	 * Executes the query without actually updating the state
@@ -424,10 +442,12 @@ SELECT COUNT(*) as rowCount FROM (${this.text.trim()})
 			(this.#executeQuery);
 
 		this.#debugStyled('length query text', '\n' + lengthQuery, 'font-family: monospace;');
-
+		const before = performance.now();
 		const resolved = resolveMaybePromise(
 			/** @returns {MaybePromise<Query<RowType>>} */
 			(lengthResult, isPromise) => {
+				const after = performance.now();
+				this.#lengthQueryTime = after - before;
 				this.#length = lengthResult[0].rowCount;
 				this.#sharedLengthPromise.resolve(this);
 				if (isPromise) {
@@ -485,9 +505,11 @@ DESCRIBE ${this.text.trim()}
 		const typedRunner =
 			/** @type {import('../types.js').Runner<import('../../types/duckdb-wellknown.js').DescribeResultRow>} */
 			(this.#executeQuery);
-
+		const before = performance.now();
 		const resolved = resolveMaybePromise(
 			(description, isPromise) => {
+				const after = performance.now();
+				this.#columnsQueryTime = after - before;
 				// Update inner value
 				this.#columns = description;
 				// Resolve store
@@ -828,9 +850,10 @@ DESCRIBE ${this.text.trim()}
 			};
 		} else if (optsOrId) {
 			opts = optsOrId;
+			if (!opts.id) opts.id = queryHash + '-' + Math.random().toString(36).substring(0, 4);
 		} else {
 			opts = {
-				id: queryHash
+				id: queryHash + '-' + Math.random().toString(36).substring(0, 4)
 			};
 		}
 		if (!('autoScore' in opts)) {
@@ -869,6 +892,7 @@ DESCRIBE ${this.text.trim()}
 
 		Query.#constructing = true;
 		const output = new Query(query, executeQuery, opts);
+		Query.#globalEmit('queryCreated', { raw: output, proxied: output.value });
 		if (!opts.disableCache) {
 			Query.#addToCache(output);
 			Query.#cacheCleanup();
@@ -1043,6 +1067,7 @@ DESCRIBE ${this.text.trim()}
 			if (!Array.isArray(knownColumns))
 				throw new Error(`Expected knownColumns to be an array`, { cause: knownColumns });
 			this.#columns = knownColumns;
+			this.#sharedColumnsPromise.resolve(this);
 		} else {
 			resolveMaybePromise(
 				() => {
@@ -1191,9 +1216,9 @@ DESCRIBE ${this.text.trim()}
 				const exactMatch = taggedSql`CASE WHEN lower("${col.trim()}") = lower('${escapedSearchTerm}') THEN 2 ELSE 0 END`;
 				const similarity = taggedSql`jaccard(lower('${escapedSearchTerm}'), lower("${col}"))`;
 				const exactSubMatch =
-					// escapedSearchTerm.length >= 4
-					taggedSql`CASE WHEN lower("${col.trim()}") LIKE lower('%${escapedSearchTerm.split(' ').join('%')}%') THEN 1 ELSE 0 END`;
-				// : taggedSql`0`;
+					escapedSearchTerm.length >= 1
+						? taggedSql`CASE WHEN lower("${col.trim()}") LIKE lower('%${escapedSearchTerm.split(' ').join('%')}%') THEN 1 ELSE 0 END`
+						: taggedSql`0`;
 				return taggedSql`GREATEST((${exactMatch}), (${similarity}), (${exactSubMatch}))`;
 			})
 			.join(',');