+++ packages/ui/core-components/src/lib/atoms/accordion/AccordionItem.svelte
@@ -5,10 +5,13 @@
 <script>
 	import * as BaseAccordion from '../shadcn/accordion';
 	export let title = '';
+
+	let className = undefined;
+	export { className as class };
 </script>
 
 {#key title}
-	<BaseAccordion.Item value={title}>
+	<BaseAccordion.Item value={title} class={className}>
 		<BaseAccordion.Trigger>
 			<slot name="title">
 				{title}