+++ packages/ui/core-components/src/lib/atoms/index.js
@@ -9,3 +9,4 @@ export * from './inputs';
 export * from './flowchart';
 export * from './grid';
 export * from './hover-card';
+export * from './query-debugger';