+++ packages/ui/core-components/src/lib/atoms/inputs/dropdown/stories/DependentDropdowns.story.svelte
@@ -39,33 +39,11 @@
 		callback: ($v) => (depQuery = $v),
 		loadGracePeriod: 2000
 	});
-
-	$: displayQueryFactory(`
-        SELECT '${$inputs?.user?.value ?? -1}'
-    `);
-	let displayQuery;
-	const displayQueryFactory = Query.createReactive({
-		execFn: slowQuery,
-		callback: ($v) => (displayQuery = $v),
-		loadGracePeriod: 2000
-	});
 </script>
 
-<div class="grid grid-cols-2">
-	<div>
-		<Dropdown defaultValue={0} name="hashtag" data={baseQuery} value="value" label="label" />
-
-		<QueryLoad let:loaded data={depQuery}>
-			<Dropdown name="user" data={loaded} value="value" label="label" multiple selectAllByDefault>
-				<DropdownOption value="All" />
-			</Dropdown>
-		</QueryLoad>
-	</div>
-	<div>
-		<QueryLoad let:loaded data={depQuery}>
-			<pre class="text-xs">{loaded.originalText}</pre>
-			<pre class="text-xs overflow-y-auto">{JSON.stringify(loaded, null, 2)}</pre>
-		</QueryLoad>
-		{displayQuery.text}
-	</div>
-</div>
+<Dropdown defaultValue={0} name="hashtag" data={baseQuery} value="value" label="label" />
+<QueryLoad let:loaded data={depQuery}>
+	<Dropdown name="user" data={loaded} value="value" label="label" multiple selectAllByDefault>
+		<DropdownOption value="All" />
+	</Dropdown>
+</QueryLoad>