+++ packages/ui/core-components/src/lib/utils.js
@@ -59,3 +59,25 @@ export const flyAndScale = (node, params = { y: -8, x: 0, start: 0.95, duration:
 		easing: cubicOut
 	};
 };
+
+/**
+ * @param {unknown} value
+ * @returns {number | undefined}
+ */
+export const toNumber = (value) => {
+	if (typeof value === 'undefined') return undefined;
+	return Number(value);
+};
+
+/**
+ * @param {unknown} value
+ * @returns {boolean | undefined}
+ */
+export const toBoolean = (value) => {
+	if (typeof value === 'undefined') return undefined;
+	if (typeof value === 'string') {
+		if (value.toLowerCase() === 'true') return true;
+		if (value.toLowerCase() === 'false') return false;
+	}
+	return Boolean(value);
+};