+++ packages/ui/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @evidence-dev/icons
 
+## 1.0.1
+
+### Patch Changes
+
+- 6cdedceb0: Remove db-orchestrator
+
 ## 1.0.0
 
 ### Major Changes